requirements.txt
certifi==2019.9.11
chardet==3.0.4
chromote==0.4.0
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.6
websocket-client==0.56.0
Werkzeug==0.16.0
Flask-Cors==3.0.8
Pillow==7.0.0
